+++ docs/en/docs/release-notes.md
@@ -5,6 +5,16 @@ hide:
 
 # Release Notes
 
+## 0.5.1
+
+### Add
+
+- `send()` as alternative to `enqueue`.
+
+### Changed
+
+- Allow `enqueue`, `enqueue_delayed` and `delay` to return the job id directly.
+
 ## 0.5.0
 
 ### Added